Add doc tests to metrics_dashboard component and extract widget types
Extract MetricKind and MetricWidget from metrics_dashboard/mod.rs into
widget.rs submodule to keep the file under 1000 lines. Add doc tests to
all 41 public methods (100% coverage), up from 4/41 (9.8%). Also fix
import ordering in chat_view/mod.rs for cargo fmt compliance.
